# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Central Global University LMS/SMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /progress:
    get:
      operationId: listProgress
      tags: [progress]
      summary: List the current student's material progress
      parameters:
        - name: courseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Material progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaterialProgress"
    post:
      operationId: recordProgress
      tags: [progress]
      summary: Record progress for a study material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialProgressInput"
      responses:
        "201":
          description: Recorded progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialProgress"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List the current student's payments
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/paypal/create-order:
    post:
      operationId: createPaypalOrder
      tags: [payments]
      summary: Create a PayPal order for a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaypalOrderInput"
      responses:
        "201":
          description: Created PayPal order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaypalOrderResponse"
  /payments/paypal/capture:
    post:
      operationId: capturePaypalOrder
      tags: [payments]
      summary: Capture an approved PayPal order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaypalCaptureInput"
      responses:
        "200":
          description: Captured payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    MaterialProgress:
      type: object
      required: [id, userId, materialId, courseId, action, completedAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        materialId: { type: integer }
        courseId: { type: integer }
        action: { type: string }
        completedAt: { type: string, format: date-time }
    MaterialProgressInput:
      type: object
      required: [materialId]
      properties:
        materialId: { type: integer }
        action: { type: string, enum: [watched, downloaded, completed] }
    PaypalOrderInput:
      type: object
      required: [courseId, returnUrl, cancelUrl]
      properties:
        courseId: { type: integer }
        returnUrl: { type: string }
        cancelUrl: { type: string }
    PaypalOrderResponse:
      type: object
      required: [paymentId, orderId, approveUrl]
      properties:
        paymentId: { type: integer }
        orderId: { type: string }
        approveUrl: { type: string }
    PaypalCaptureInput:
      type: object
      required: [orderId]
      properties:
        orderId: { type: string }
...
    Payment:
      type: object
      required: [id, userId, amount, currency, status, provider, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        courseId: { type: ["integer", "null"] }
        amount: { type: number }
        currency: { type: string }
        status: { type: string, enum: [pending, completed, failed, refunded] }
        provider: { type: string }
        reference: { type: ["string", "null"] }
        invoiceNumber: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    PaymentInput:
      type: object
      required: [amount]
      properties:
        amount: { type: number }
        courseId: { type: integer }
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }