# lib/api-zod/src/generated/types/paypalOrderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Central Global University LMS/SMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaypalOrderInput {
  courseId: number;
  returnUrl: string;
  cancelUrl: string;
}